Fix. 데이터 수정

+++ app/src/main/java/com/example/meetup/model/CategoryIdResponseModel.kt
@@ -15,5 +15,5 @@ data class Food(
     val dollarPrice: Long,
     val canadaPrice: Long,
     val image: String?,
-    val avgRate: Long,
+    val avgRate: Double,
 )
