+++ app/src/main/java/com/example/meetup/model/store/GetStoreDetailResponseModel.kt
@@ -20,7 +20,7 @@ data class GetStoreDetailStoreDto(
     var name: String,
     var description: String,
     var koreanYn: Boolean,
-    var avgRate: Long,
+    var avgRate: Double,
     var minOrderAmount: Long,
     var images: ArrayList<String>
 )
